app/src/main/java/com/example/yanghanwen/taskmanagementmonster/Message.java
package com.example.yanghanwen.taskmanagementmonster;


import java.util.Date;


/*
 *
 *  * Copyright © 2018 CMPUT301W18T08, University of Alberta - All Rights Reserved.
 *  * You may use, distribute or modify this code under terms and conditions of Code of Student Behavior at
 *  *  University of Alberta.
 *  * You can find a copy of the license in this project, otherwise please contact at
 *  *   hyang4@ualberta.ca
 *
 *
 */

/**
 * Activity current has no use
 */
public class Message {
    private final String username;
    private String message;
    private Date date;


    public Message(String username, String message, Date date){
        this.username = username;
        this.message = message;
        this.date = date;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }

    public Date getDate() {
        return date;
    }

    public void setDate(Date date) {
        this.date = date;
    }
}
